Update setup.py values; document release process
This moves the version information to fx_crash_sig/__init__.py rather
than rely on pkg_resources.

This updates the package url to the correct value and removes the
maintainer because there are a bunch of us maintaining.

This documents a release process to make it easier for all of us.