Fix AttributeErrors (#23)
This fixes an attribute error when printing out details of an unhandled
exception when using the fx-crash-sig command.

This fixes the attribute error caused by some parts of the code still
using a non-existent `__postprocess` method.

This fixes the problem where siggen now returns a Result object which
isn't serializable and so it doesn't work with json.dumps.