Fix symbolication issues (#32)
This is a big set of changes.

It looks like the structure of crash pings have changed over the last
few years. This updates the code to handle the current version. Plus it
changes the code to work with the whole crash ping rather than pieces of
it. It specifies which pieces it needs in the comments.

This gets rid of the "windows flag" in favor of looking at the
normalized_os field of the crash ping.

This creates a SYMBOLICATION_API constant and moves it to the top level.

This fixes fx-crash-sig and sample.json so those work. This also updates
the examples in README.md.

This updates setup.py--this is a Python 3 library.

Finally, it fixes the issue where the output of symbolication puts the
debug_file (e.g. xul.pdb) in the module field of frames. Signature
generation expects that field to be the filename (e.g. xul.dll). I claim
that's a bug in the symbolication API code, but until it's fixed there,
we have to fix it in fx-crash-sig.